compiler warnings

+++ src/main.c
@@ -716,7 +716,7 @@ sendSignal(void)
 	    !(opt_send_signal == 0 && errno == EPERM)) {
 	    fprintf(stderr, "%s: ERROR: Could not send ", appname);
 	    fprintf(stderr, "signal %d to process %d: %s\n",
-		opt_send_signal, pid, xstrerror());
+		opt_send_signal, (int) pid, xstrerror());
 	    exit(1);
 	}
     } else {
